+++ src/main/java/com/airbnl/hotelservice/service/interfaces/IRoomService.java
@@ -7,5 +7,5 @@ import java.util.List;
 
 public interface IRoomService {
     List<Room> getRoomsByHotelId(Hotel hotel);
-    List<Room> getRoomsByCountryAndRoomTypeIds(int countryId,int roomTypeId);
+    List<Room> getRoomsByCountryAndRoomTypeIds(long countryId,long roomTypeId);
 }
